+++ tests/test_email_notifications.py
@@ -14,7 +14,7 @@ vcr = VCR(decode_compressed_response=True)
 class EmailNotificationsTestCase(TestCase):
     @freeze_time("2018-01-01")
     def test_2sa_required_email_notification(self):
-        base_dir = os.path.normpath(f"tests/fixtures/Photos/{inspect.stack()[0][3]}")
+        base_dir = os.path.normpath(f"tests/fixtures/{inspect.stack()[0][3]}")
         if os.path.exists(base_dir):
             shutil.rmtree(base_dir)
         os.makedirs(base_dir)
@@ -64,7 +64,7 @@ class EmailNotificationsTestCase(TestCase):
 
     @freeze_time("2018-01-01")
     def test_2sa_notification_without_smtp_login_and_tls(self):
-        base_dir = os.path.normpath(f"tests/fixtures/Photos/{inspect.stack()[0][3]}")
+        base_dir = os.path.normpath(f"tests/fixtures/{inspect.stack()[0][3]}")
         if os.path.exists(base_dir):
             shutil.rmtree(base_dir)
         os.makedirs(base_dir)
@@ -109,7 +109,7 @@ class EmailNotificationsTestCase(TestCase):
 
     @freeze_time("2018-01-01")
     def test_2sa_required_notification_script(self):
-        base_dir = os.path.normpath(f"tests/fixtures/Photos/{inspect.stack()[0][3]}")
+        base_dir = os.path.normpath(f"tests/fixtures/{inspect.stack()[0][3]}")
         if os.path.exists(base_dir):
             shutil.rmtree(base_dir)
         os.makedirs(base_dir)
@@ -139,7 +139,7 @@ class EmailNotificationsTestCase(TestCase):
 
     @freeze_time("2018-01-01")
     def test_2sa_required_email_notification_from(self):
-        base_dir = os.path.normpath(f"tests/fixtures/Photos/{inspect.stack()[0][3]}")
+        base_dir = os.path.normpath(f"tests/fixtures/{inspect.stack()[0][3]}")
         if os.path.exists(base_dir):
             shutil.rmtree(base_dir)
         os.makedirs(base_dir)