+++ tests/test_listing_recent_photos.py
@@ -15,7 +15,7 @@ vcr = VCR(decode_compressed_response=True)
 class ListingRecentPhotosTestCase(TestCase):
 
     def test_listing_recent_photos(self):
-        base_dir = os.path.normpath(f"tests/fixtures/Photos/{inspect.stack()[0][3]}")
+        base_dir = os.path.normpath(f"tests/fixtures/{inspect.stack()[0][3]}")
         if os.path.exists(base_dir):
             shutil.rmtree(base_dir)
         os.makedirs(base_dir)
@@ -79,7 +79,7 @@ class ListingRecentPhotosTestCase(TestCase):
         assert sum(1 for _ in files_in_result) == 0
 
     def test_listing_photos_does_not_create_folders(self):
-        base_dir = os.path.normpath(f"tests/fixtures/Photos/{inspect.stack()[0][3]}")
+        base_dir = os.path.normpath(f"tests/fixtures/{inspect.stack()[0][3]}")
         if os.path.exists(base_dir):
             shutil.rmtree(base_dir)
         os.makedirs(base_dir)
@@ -124,7 +124,7 @@ class ListingRecentPhotosTestCase(TestCase):
         assert sum(1 for _ in files_in_result) == 0
 
     def test_listing_recent_photos_with_missing_filenameEnc(self):
-        base_dir = os.path.normpath(f"tests/fixtures/Photos/{inspect.stack()[0][3]}")
+        base_dir = os.path.normpath(f"tests/fixtures/{inspect.stack()[0][3]}")
         if os.path.exists(base_dir):
             shutil.rmtree(base_dir)
         os.makedirs(base_dir)
@@ -190,7 +190,7 @@ class ListingRecentPhotosTestCase(TestCase):
     # This was used to solve the missing filenameEnc error. I found
     # another case where it might crash. (Maybe Apple changes the downloadURL key)
     def test_listing_recent_photos_with_missing_downloadURL(self):
-        base_dir = os.path.normpath(f"tests/fixtures/Photos/{inspect.stack()[0][3]}")
+        base_dir = os.path.normpath(f"tests/fixtures/{inspect.stack()[0][3]}")
         if os.path.exists(base_dir):
             shutil.rmtree(base_dir)
         os.makedirs(base_dir)